+++ src/main/kotlin/com/mdrobnak/lalrpop/psi/ext/LpGrammarParam.kt
@@ -1,0 +1,6 @@
+package com.mdrobnak.lalrpop.psi.ext
+
+import com.mdrobnak.lalrpop.psi.LpGrammarParam
+
+val LpGrammarParam.name: String
+    get() = this.firstChild.text
