+++ src/main/kotlin/com/mdrobnak/lalrpop/psi/ext/LpNonterminalName.kt
@@ -3,10 +3,11 @@ package com.mdrobnak.lalrpop.psi.ext
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
-import com.mdrobnak.lalrpop.psi.LpElementFactory
-import com.mdrobnak.lalrpop.psi.LpElementTypes
-import com.mdrobnak.lalrpop.psi.LpNonterminalName
+import com.mdrobnak.lalrpop.psi.*
 import com.mdrobnak.lalrpop.psi.impl.LpNonterminalImpl
+
+val LpNonterminalName.nonterminalParent: LpNonterminal
+    get() = this.parent as LpNonterminal
 
 abstract class LpNonterminalNameMixin(node: ASTNode) : ASTWrapperPsiElement(node), LpNonterminalName {
     override fun getNameIdentifier(): PsiElement {
