+++ src/main/kotlin/com/mdrobnak/lalrpop/psi/ext/LpRustReference.kt
@@ -2,10 +2,20 @@ package com.mdrobnak.lalrpop.psi.ext
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
+import com.mdrobnak.lalrpop.psi.LpElementTypes
 import com.mdrobnak.lalrpop.psi.LpRustReference
 import com.mdrobnak.lalrpop.psi.NonterminalGenericArgument
-import com.mdrobnak.lalrpop.psi.util.isRefMut
-import com.mdrobnak.lalrpop.psi.util.lifetimeOrInfer
+import org.rust.lang.core.psi.ext.childrenWithLeaves
+import org.rust.lang.core.psi.ext.elementType
+
+val LpRustReference.lifetime: String?
+    get() = this.childrenWithLeaves.find { it.elementType == LpElementTypes.LIFETIME }?.text
+
+val LpRustReference.lifetimeOrInfer: String
+    get() = this.lifetime ?: "'_"
+
+val LpRustReference.isRefMut: Boolean
+    get() = this.childrenWithLeaves.find { it.elementType == LpElementTypes.MUT } != null
 
 abstract class LpRustReferenceMixin(node: ASTNode) : ASTWrapperPsiElement(node), LpRustReference {
     override fun resolveType(arguments: List<NonterminalGenericArgument>): String {
